Refs #24682: Storyblok - move breakpoints calculation into BaseImage and leave only two image variants on input
